Full refactor to make testing in future easier and more robust
