+++ pkg/github/webhooks.go
@@ -1,6 +1,7 @@
 package github
 
 import (
+	"context"
 	"net/http"
 
 	"github.com/Spazzy757/paul/pkg/helpers"
@@ -8,7 +9,11 @@ import (
 )
 
 //IncomingWebhook handles an incoming webhook request
-func IncomingWebhook(r *http.Request) error {
+func IncomingWebhook(
+	ctx context.Context,
+	r *http.Request,
+	client *github.Client,
+) error {
 	// handle authentication
 	secret_key := helpers.GetEnv("SECRET_KEY", "")
 	payload, validationErr := github.ValidatePayload(r, []byte(secret_key))
@@ -22,9 +27,9 @@ func IncomingWebhook(r *http.Request) error {
 	var err error
 	switch e := event.(type) {
 	case *github.IssueCommentEvent:
-		err = IssueCommentHandler(e)
+		err = IssueCommentHandler(ctx, e, client)
 	case *github.PullRequestEvent:
-		err = PullRequestHandler(e)
+		err = PullRequestHandler(ctx, e, client)
 	default:
 		break
 	}
